Fixed bug#02.

+++ include/sum.h
@@ -1,4 +1,4 @@
 #ifndef __sum__
 #define __sum__
-int sum();
+float sum();
 #endif
