+++ src/sum.c
@@ -1,8 +1,8 @@
 #include "sum.h"
 #include "operand.h" 
-int sum(int argc, char** argv)
+float sum(int argc, char** argv)
 {
-	int res=0;
+	float res=0;
 	res=sum_of_operand(argc, argv);
 	return res;
